Reduce the unnecessary boilerplate around receiving Playlist informations
